fix thanh toan vnpay

+++ app/Http/Controllers/HoaDonController.php
@@ -30,10 +30,7 @@ class HoaDonController extends Controller
 
     // thanh toán online
 
-    
-
-    public function returnOnline(Request $request)
-    {
+    public function return() {
         $url = session('url_prev','/');
         $total = session('tong_tien');
         session()->forget('tong_tien');
@@ -43,9 +40,72 @@ class HoaDonController extends Controller
             $bill['ngay_lap_hd'] = Carbon::now();
             $bill['trang_thai'] = true;
             $bill_id = HoaDon::insertGetId($bill);
+
+            $gio_hangs = GioHang::where('tai_khoans_id',Auth::user()->id)->get();
+            $loi_nhuan = 0;
+            $tongtien = 0;
+            foreach($gio_hangs as $gio_hang) {
+                $giaban = $gio_hang->chiTietSanPham->sanpham->gia_ban*(100-$gio_hang->chiTietSanPham->sanpham->giam_gia)/100;
+                $giagoc = $gio_hang->chiTietSanPham->sanPham->gia_goc;
+                $bill_detail = new ChiTietHoaDon();
+                $bill_detail->hoa_dons_id = $bill_id;
+                $bill_detail->chi_tiet_san_phams_id = $gio_hang->chi_tiet_san_phams_id;
+                $bill_detail->so_luong = $gio_hang->so_luong; 
+                $bill_detail->gia_goc = $giagoc;
+                $bill_detail->gia_ban = $giaban;
+                $update_ctsp = ChiTietSanPham::find($content->id);
+                $update_ctsp->so_luong -= $content->qty;
+                $update_ctsp->save();
+                $bill_detail->save();
+                $tongtien += ($giaban*$gio_hang->so_luong);
+                $loi_nhuan += (($giaban - $giagoc)*$gio_hang->so_luong);
+            }
             $add_loi_nhuan = HoaDon::find($bill_id);
-            $add_loi_nhuan->tong_tien = $total;
+            $add_loi_nhuan->loi_nhuan = $loi_nhuan;
+            $add_loi_nhuan->tong_tien = $tongtien;
             $add_loi_nhuan->save();
+
+            $delete = GioHang::where('tai_khoans_id', Auth::user()->id)->delete();
+            return redirect()->route('ordersuccess');
+        }
+        session()->forget('url_prev');
+        return redirect($url)->with('errors' ,'Lỗi trong quá trình thanh toán phí dịch vụ');
+    }
+
+    public function returnNotAcc($id)
+    {
+        $url = session('url_prev','/');
+        $total = session('tong_tien');
+        session()->forget('tong_tien');
+        if($request->vnp_ResponseCode == "00") {
+            $bill = array();
+            $bill['tai_khoans_id'] = $id;
+            $bill['ngay_lap_hd'] = Carbon::now();
+            $bill['trang_thai'] = true;
+            $bill_id = HoaDon::insertGetId($bill);
+
+            $contents = Cart::content();
+            $loi_nhuan = 0;
+            $tongtien = 0;
+            foreach($contents as $content) {
+                $bill_detail = new ChiTietHoaDon();
+                $bill_detail->hoa_dons_id = $bill_id;
+                $bill_detail->chi_tiet_san_phams_id = $content->id;
+                $bill_detail->so_luong = $content->qty; 
+                $update_ctsp = ChiTietSanPham::find($content->id);
+                $update_ctsp->so_luong -= $content->qty;
+                $bill_detail->gia_goc = $update_ctsp->sanPham->gia_goc;
+                $bill_detail->gia_ban = $content->price;
+                $tongtien += ($content->price*$content->qty);
+                $loi_nhuan += (($content->price - $update_ctsp->sanPham->gia_goc)*$content->qty);
+                $update_ctsp->save();
+                $bill_detail->save();
+            }
+            $add_loi_nhuan = HoaDon::find($bill_id);
+            $add_loi_nhuan->loi_nhuan = $loi_nhuan;
+            $add_loi_nhuan->tong_tien = $tongtien;
+            $add_loi_nhuan->save();
+
             Cart::destroy();
             return redirect()->route('ordersuccess');
         }
@@ -56,6 +116,7 @@ class HoaDonController extends Controller
     // end
 
     public function create(Request $request){
+        //thanh toán vnpay
         if(!empty($request->payment))
         {
         session(['tong_tien' => $request->online]);
@@ -112,6 +173,7 @@ class HoaDonController extends Controller
         }
         return redirect($vnp_Url);
         }
+        //ket thuc thanh toan vnpay
         if(Auth::check() and Auth::user()->admin != 1) {
             if(Auth::user()->dia_chi == null) {
                 $request->validate([
@@ -133,6 +195,64 @@ class HoaDonController extends Controller
                 $new_update->so_dien_thoai = $request->so_dien_thoai;
                 $new_update->save();
             }
+            //thanh toán vnpay
+        if(!empty($request->payment))
+        {
+        session(['tong_tien' => $request->online]);
+        session(['cost_id' => $request->id]);
+        session(['url_prev' => url()->previous()]);
+        $vnp_TmnCode = "UDOPNWS1"; //Mã website tại VNPAY 
+        $vnp_HashSecret = "EBAHADUGCOEWYXCMYZRMTMLSHGKNRPBN"; //Chuỗi bí mật
+        $vnp_Url = "http://sandbox.vnpayment.vn/paymentv2/vpcpay.html";
+        $vnp_Returnurl = "http://localhost:8000/return-vnpay";
+        $vnp_TxnRef = date("YmdHis"); //Mã đơn hàng. Trong thực tế Merchant cần insert đơn hàng vào DB và gửi mã này sang VNPAY
+        $vnp_OrderInfo = "Thanh toán hóa đơn phí dich vụ";
+        $vnp_OrderType = 'billpayment';
+        $vnp_Amount = $request->online * 100;
+        $vnp_Locale = 'vn';
+        $vnp_IpAddr = request()->ip();
+
+        $inputData = array(
+            "vnp_Version" => "2.0.0",
+            "vnp_TmnCode" => $vnp_TmnCode,
+            "vnp_Amount" => $vnp_Amount,
+            "vnp_Command" => "pay",
+            "vnp_CreateDate" => date('YmdHis'),
+            "vnp_CurrCode" => "VND",
+            "vnp_IpAddr" => $vnp_IpAddr,
+            "vnp_Locale" => $vnp_Locale,
+            "vnp_OrderInfo" => $vnp_OrderInfo,
+            "vnp_OrderType" => $vnp_OrderType,
+            "vnp_ReturnUrl" => $vnp_Returnurl,
+            "vnp_TxnRef" => $vnp_TxnRef,
+        );
+        $vnp_BankCode = 'ncb';
+        if (isset($vnp_BankCode) && $vnp_BankCode != "") {
+            $inputData['vnp_BankCode'] = $vnp_BankCode;
+        }
+        ksort($inputData);
+        $query = "";
+        $i = 0;
+        $hashdata = "";
+        foreach ($inputData as $key => $value) {
+            if ($i == 1) {
+                $hashdata .= '&' . $key . "=" . $value;
+            } else {
+                $hashdata .= $key . "=" . $value;
+                $i = 1;
+            }
+            $query .= urlencode($key) . "=" . urlencode($value) . '&';
+        }
+
+        $vnp_Url = $vnp_Url . "?" . $query;
+        if (isset($vnp_HashSecret)) {
+           // $vnpSecureHash = md5($vnp_HashSecret . $hashdata);
+            $vnpSecureHash = hash('sha256', $vnp_HashSecret . $hashdata);
+            $vnp_Url .= 'vnp_SecureHashType=SHA256&vnp_SecureHash=' . $vnpSecureHash;
+        }
+        return redirect($vnp_Url);
+        }
+        //ket thuc thanh toan vnpay
             $bill = array();
             $bill['tai_khoans_id'] = Auth::user()->id;
             $bill['ngay_lap_hd'] = Carbon::now();
@@ -162,6 +282,8 @@ class HoaDonController extends Controller
             $add_loi_nhuan->loi_nhuan = $loi_nhuan;
             $add_loi_nhuan->tong_tien = $tongtien;
             $add_loi_nhuan->save();
+
+            $delete = GioHang::where('tai_khoans_id', Auth::user()->id)->delete();
         }
         else {
             $request->validate([
@@ -183,17 +305,132 @@ class HoaDonController extends Controller
                 $account['admin'] = false;
                 $account_id = TaiKhoan::insertGetId($account);
 
-                $total = Cart::subtotal();
+                    //thanh toán vnpay
+        if(!empty($request->payment))
+        {
+        session(['tong_tien' => $request->online]);
+        session(['cost_id' => $request->id]);
+        session(['url_prev' => url()->previous()]);
+        $vnp_TmnCode = "UDOPNWS1"; //Mã website tại VNPAY 
+        $vnp_HashSecret = "EBAHADUGCOEWYXCMYZRMTMLSHGKNRPBN"; //Chuỗi bí mật
+        $vnp_Url = "http://sandbox.vnpayment.vn/paymentv2/vpcpay.html";
+        $vnp_Returnurl = "http://localhost:8000/return-vnpay-not-acc/{{$account_id}}";
+        $vnp_TxnRef = date("YmdHis"); //Mã đơn hàng. Trong thực tế Merchant cần insert đơn hàng vào DB và gửi mã này sang VNPAY
+        $vnp_OrderInfo = "Thanh toán hóa đơn phí dich vụ";
+        $vnp_OrderType = 'billpayment';
+        $vnp_Amount = $request->online * 100;
+        $vnp_Locale = 'vn';
+        $vnp_IpAddr = request()->ip();
+
+        $inputData = array(
+            "vnp_Version" => "2.0.0",
+            "vnp_TmnCode" => $vnp_TmnCode,
+            "vnp_Amount" => $vnp_Amount,
+            "vnp_Command" => "pay",
+            "vnp_CreateDate" => date('YmdHis'),
+            "vnp_CurrCode" => "VND",
+            "vnp_IpAddr" => $vnp_IpAddr,
+            "vnp_Locale" => $vnp_Locale,
+            "vnp_OrderInfo" => $vnp_OrderInfo,
+            "vnp_OrderType" => $vnp_OrderType,
+            "vnp_ReturnUrl" => $vnp_Returnurl,
+            "vnp_TxnRef" => $vnp_TxnRef,
+        );
+        $vnp_BankCode = 'ncb';
+        if (isset($vnp_BankCode) && $vnp_BankCode != "") {
+            $inputData['vnp_BankCode'] = $vnp_BankCode;
+        }
+        ksort($inputData);
+        $query = "";
+        $i = 0;
+        $hashdata = "";
+        foreach ($inputData as $key => $value) {
+            if ($i == 1) {
+                $hashdata .= '&' . $key . "=" . $value;
+            } else {
+                $hashdata .= $key . "=" . $value;
+                $i = 1;
+            }
+            $query .= urlencode($key) . "=" . urlencode($value) . '&';
+        }
+
+        $vnp_Url = $vnp_Url . "?" . $query;
+        if (isset($vnp_HashSecret)) {
+           // $vnpSecureHash = md5($vnp_HashSecret . $hashdata);
+            $vnpSecureHash = hash('sha256', $vnp_HashSecret . $hashdata);
+            $vnp_Url .= 'vnp_SecureHashType=SHA256&vnp_SecureHash=' . $vnpSecureHash;
+        }
+        return redirect($vnp_Url);
+        }
+        //ket thuc thanh toan vnpay
+    
                 $bill = array();
                 $bill['tai_khoans_id'] = $account_id;
                 $bill['ngay_lap_hd'] = Carbon::now();
-                $bill['tong_tien'] = $total;
                 $bill['trang_thai'] = true;
                 $bill_id = HoaDon::insertGetId($bill);
             }
-
-
             else { 
+
+                //thanh toán vnpay
+        if(!empty($request->payment))
+        {
+        session(['tong_tien' => $request->online]);
+        session(['cost_id' => $request->id]);
+        session(['url_prev' => url()->previous()]);
+        $vnp_TmnCode = "UDOPNWS1"; //Mã website tại VNPAY 
+        $vnp_HashSecret = "EBAHADUGCOEWYXCMYZRMTMLSHGKNRPBN"; //Chuỗi bí mật
+        $vnp_Url = "http://sandbox.vnpayment.vn/paymentv2/vpcpay.html";
+        $vnp_Returnurl = "http://localhost:8000/return-vnpay-not-acc/{{$check_account->id}}";
+        $vnp_TxnRef = date("YmdHis"); //Mã đơn hàng. Trong thực tế Merchant cần insert đơn hàng vào DB và gửi mã này sang VNPAY
+        $vnp_OrderInfo = "Thanh toán hóa đơn phí dich vụ";
+        $vnp_OrderType = 'billpayment';
+        $vnp_Amount = $request->online * 100;
+        $vnp_Locale = 'vn';
+        $vnp_IpAddr = request()->ip();
+
+        $inputData = array(
+            "vnp_Version" => "2.0.0",
+            "vnp_TmnCode" => $vnp_TmnCode,
+            "vnp_Amount" => $vnp_Amount,
+            "vnp_Command" => "pay",
+            "vnp_CreateDate" => date('YmdHis'),
+            "vnp_CurrCode" => "VND",
+            "vnp_IpAddr" => $vnp_IpAddr,
+            "vnp_Locale" => $vnp_Locale,
+            "vnp_OrderInfo" => $vnp_OrderInfo,
+            "vnp_OrderType" => $vnp_OrderType,
+            "vnp_ReturnUrl" => $vnp_Returnurl,
+            "vnp_TxnRef" => $vnp_TxnRef,
+        );
+        $vnp_BankCode = 'ncb';
+        if (isset($vnp_BankCode) && $vnp_BankCode != "") {
+            $inputData['vnp_BankCode'] = $vnp_BankCode;
+        }
+        ksort($inputData);
+        $query = "";
+        $i = 0;
+        $hashdata = "";
+        foreach ($inputData as $key => $value) {
+            if ($i == 1) {
+                $hashdata .= '&' . $key . "=" . $value;
+            } else {
+                $hashdata .= $key . "=" . $value;
+                $i = 1;
+            }
+            $query .= urlencode($key) . "=" . urlencode($value) . '&';
+        }
+
+        $vnp_Url = $vnp_Url . "?" . $query;
+        if (isset($vnp_HashSecret)) {
+           // $vnpSecureHash = md5($vnp_HashSecret . $hashdata);
+            $vnpSecureHash = hash('sha256', $vnp_HashSecret . $hashdata);
+            $vnp_Url .= 'vnp_SecureHashType=SHA256&vnp_SecureHash=' . $vnpSecureHash;
+        }
+        return redirect($vnp_Url);
+        }
+        //ket thuc thanh toan vnpay
+
                 $bill = array();
                 $bill['tai_khoans_id'] = $check_account->id;
                 $bill['ngay_lap_hd'] = Carbon::now();
